# internal/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Arduino-App-Cli
  version: 0.1.0
servers:
- description: local server
  url: http://localhost:8800
...
paths:
...
  /v1/bricks:
    get:
      description: Returns all the existing bricks. Bricks that are ready to use are
        marked as installed.
      operationId: getBricks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrickListResult'
          description: Successful response
        "500":
          $ref: '#/components/responses/InternalServerError'
      summary: Get a list of available bricks
      tags:
      - Brick
...
  /v1/models/ei/projects/{projectID}:
    put:
      description: Download and install a custom Edge Impulse AI model using the provided
        project API key.
      operationId: installEIModel
      parameters:
      - description: Edge Impulse project ID
        in: path
        name: projectID
        required: true
        schema:
          description: Edge Impulse project ID
          example: 123456
          type: integer
      - description: Edge Impulse project API key
        in: header
        name: x-api-key
        required: true
        schema:
          description: Edge Impulse project API key
          example: your_edge_impulse_api_token
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                impulse_id:
                  description: Edge Impulse impulse ID
                  example: 1
                  type: integer
              required:
              - impulse_id
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelItem'
          description: Successful response
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalServerError'
        "507":
          $ref: '#/components/responses/InsufficientStorage'
      summary: Download and install a custom Edge Impulse AI model
      tags:
      - AIModels
...
components:
  responses:
    BadRequest:
      content:
        application/json:
          example:
            details: The request is invalid or missing required parameters.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Bad Request
...
    Forbidden:
      content:
        application/json:
          example:
            details: You do not have permission to access this resource.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Forbidden
    InsufficientStorage:
      content:
        application/json:
          example:
            details: Insufficient storage to complete the request.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Insufficient Storage
    InternalServerError:
      content:
        application/json:
          example:
            details: An unexpected error occurred.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Internal Server Error
...
    Unauthorized:
      content:
        application/json:
          example:
            details: Unauthorized access to the resource.
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Unauthorized
  schemas:
...
    AIModelItem:
      properties:
        brick_ids:
          items:
            type: string
          nullable: true
          type: array
        description:
          type: string
        handler:
          description: handler that manages the model, as the listing reports it (hf-handler,
            ei-handler, ai-hub-handler)
          type: string
        id:
          example: bGxhbWFjcHA6Z2VtbWEtMy0xYi1pdC1RNF8w
          type: string
        id_decoded:
          example: llamacpp:gemma-3-1b-it-Q4_0
          type: string
        metadata:
          additionalProperties:
            type: string
          type: object
        name:
          type: string
        origin:
          $ref: '#/components/schemas/ModelOrigin'
        preinstalled:
          description: 'ships with the board image: nothing to download, cannot be
            deleted'
          type: boolean
        runner:
          type: string
        size_bytes:
          description: size in bytes, converted from MiB rounded to two decimals (approximate
            to ~5 KiB); on disk when installed, declared otherwise; omitted when unknown
          minimum: 0
          nullable: true
          type: integer
        status:
          $ref: '#/components/schemas/ModelStatus'
      type: object
...
    BrickListItem:
      properties:
        author:
          type: string
        category:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        require_model:
          type: boolean
        status:
          type: string
      type: object
    BrickListResult:
      properties:
        bricks:
          items:
            $ref: '#/components/schemas/BrickListItem'
          nullable: true
          type: array
      type: object
...
    ErrorResponse:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
...
    ModelOrigin:
      description: 'Where the model came from: "curated" is declared by the internal
        model list and installs from its id alone, "user" was downloaded from a source
        the caller supplied (a Hugging Face link, an Edge Impulse project) and needs
        that source again.'
      enum:
      - curated
      - user
      type: string
      uniqueItems: true
    ModelStatus:
      description: Model status
      enum:
      - installed
      - not-installed
      - downloading
      type: string
      uniqueItems: true
...